requirements.txt
Flask
requests
firebase-admin
pycryptodome
opencv-python
numpy
scipy
insightface
